Conditional rendering by notification type  - refactor files components/notification*, split them up
