+++ frontend/src/components/mainView.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {User} from 'types';
 import {NotificationIterator, NotificationSet} from "notificationSet";
-import {NotificationContainer} from 'components/notificationView'
+import {NotificationContainer} from 'components/notificationContainer'
 import TopBarView from 'components/topBarView'
 import LoadingController from "../loadingController";
 
@@ -25,8 +25,6 @@ export default class MainView extends React.Component<MainViewProps, MainViewSta
     }
 
     render(): React.ReactNode {
-        console.log(JSON.stringify(this.props.notifications));
-        console.log(this.props.notifications);
         return <div id="mainView">
             <TopBarView user={this.props.user} />
             <NotificationContainer notifications={this.props.notifications} loadingController={this.props.loadingController} />
